requirements.txt
docker-py==1.7.0
click
